Add well-known label to InferenceGraphs (#463)
Implementation of authentication for InferenceGraphs using Serverless mode will use ServiceMesh with Authorino (like InferenceServices case). For this, an Istio AuthorizationPolicy is applied mesh-wide. The AuthorizationPolicy uses a static label selector that is expected to be present on all Pods that require to be protected.

Currently, the InferenceGraph does not have a static label that can be used.

This PR is adding the `serving.kserve.io/kind: InferenceGraph` to pods that belong to InferenceGraph to properly protect them per user request.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>