Use leader election for better compatibility with in-place upgrades (#492)
* Revert "Better compatibility with in-place upgrades (#491)"

This reverts commit 2a917eafe8ec716254423c9079f05aaa7f59ab0f.

* Use leader election for better compatibility with in-place upgrades

The previous commit reverts 2a917e, which changed the configuration of the Deployment for the manager to use `Recreate` strategy. Such change was for better compatibility for in-place upgrades (see the commit message of the reverted one).

Instead of changing the Deployment strategy, this commit is enabling leader election in the manager. The leader election also solves the issues mentioned in commit 2a917e. Also, not changing the Deployment strategy works better with odh-operator.

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>

---------

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>