fix(config): reworks config to be immutable (#711)
Previous implementation did not account for race conditions when
multiple concurrent reconciles are taking place (when
`MaxConcurrentReconciles` is set to `> 1`).

This PR reworks how config loading is handled by ensuring that each
reconcile gets a consistent (local) instance of the configuration.

This eliminates relying on shared state that could lead to
inconsistencies and makes the code easier to reason about.

Signed-off-by: Bartosz Majsak <bartosz.majsak@gmail.com>