Avoid service account errors, propagate serviceaccount spec for model artifacts attachment (#880)
These error logs are a product of the order in which reconciliation is performed for multi node llmisvcs. Service account reconciliation is performed as the first part of multi node llmisvc reconciliation. In the case of a desired single node llmisvc, the service account will be deleted if it exists, or not created. The same thing applies in reverse for single node reconciliation when multi-node is desired. After the service account is reconciled the workload reconciliation will continue and utilize the llmisvc spec which contains the nonexistent service account. When the expected pod spec is created for the llmisvc as a part of this continued reconciliation it will attach model artifacts to the pod spec. Given that the credential builder used for model artifact attachment relies completely on credentials configured via service accounts, if the service account set in the pod spec does not exist the credential builder will log the above error.

To suppress this error, a new method CreateSecretVolumeAndEnvFromServiceAccount has been added to the credential builder. This method takes a service account spec as input, thus avoiding the need to fetch the service account. This allows us to pass the expected service account spec directly from llmisvc reconciliation to the attach model artifacts functionality. In scenarios where no service account is set in the llmisvc template spec, we will pass the service account as nil and the default service account from the same namespace as the llmisvc will be utilized.

Signed-off-by: Brett Thompson <196701379+brettmthompson@users.noreply.github.com>